+++ lrtable/src/lib/pager.rs
@@ -311,7 +311,7 @@ fn gc<StorageT: Eq + Hash + PrimInt>
     // First of all, do a simple pass over all states. All state indexes reachable from the
     // start state will be inserted into the 'seen' set.
     let mut todo = HashSet::new();
-    todo.insert(StIdx::from(0 as u32));
+    todo.insert(StIdx::from(0u32));
     let mut seen = HashSet::new();
     while !todo.is_empty() {
         // XXX This is the clumsy way we're forced to do what we'd prefer to be:
@@ -432,12 +432,12 @@ mod test {
         assert_eq!(sg.all_states_len(), 10);
         assert_eq!(sg.all_edges_len(), 10);
 
-        assert_eq!(sg.closed_state(StIdx::from(0 as u32)).items.len(), 3);
-        state_exists(&grm, &sg.closed_state(StIdx::from(0 as u32)), "^", 0, 0, vec!["$"]);
-        state_exists(&grm, &sg.closed_state(StIdx::from(0 as u32)), "S", 0, 0, vec!["$", "b"]);
-        state_exists(&grm, &sg.closed_state(StIdx::from(0 as u32)), "S", 1, 0, vec!["$", "b"]);
+        assert_eq!(sg.closed_state(StIdx::from(0u32)).items.len(), 3);
+        state_exists(&grm, &sg.closed_state(StIdx::from(0u32)), "^", 0, 0, vec!["$"]);
+        state_exists(&grm, &sg.closed_state(StIdx::from(0u32)), "S", 0, 0, vec!["$", "b"]);
+        state_exists(&grm, &sg.closed_state(StIdx::from(0u32)), "S", 1, 0, vec!["$", "b"]);
 
-        let s1 = sg.edge(StIdx::from(0 as u32), Symbol::Nonterm(grm.nonterm_idx("S").unwrap())).unwrap();
+        let s1 = sg.edge(StIdx::from(0u32), Symbol::Nonterm(grm.nonterm_idx("S").unwrap())).unwrap();
         assert_eq!(sg.closed_state(s1).items.len(), 2);
         state_exists(&grm, &sg.closed_state(s1), "^", 0, 1, vec!["$"]);
         state_exists(&grm, &sg.closed_state(s1), "S", 0, 1, vec!["$", "b"]);
@@ -446,7 +446,7 @@ mod test {
         assert_eq!(sg.closed_state(s2).items.len(), 1);
         state_exists(&grm, &sg.closed_state(s2), "S", 0, 2, vec!["$", "b"]);
 
-        let s3 = sg.edge(StIdx::from(0 as u32), Symbol::Term(grm.term_idx("b").unwrap())).unwrap();
+        let s3 = sg.edge(StIdx::from(0u32), Symbol::Term(grm.term_idx("b").unwrap())).unwrap();
         assert_eq!(sg.closed_state(s3).items.len(), 4);
         state_exists(&grm, &sg.closed_state(s3), "S", 1, 1, vec!["$", "b", "c"]);
         state_exists(&grm, &sg.closed_state(s3), "A", 0, 0, vec!["a"]);
@@ -508,16 +508,16 @@ mod test {
         assert_eq!(sg.all_edges_len(), 27);
 
         // State 0
-        assert_eq!(sg.closed_state(StIdx::from(0 as u32)).items.len(), 7);
-        state_exists(&grm, &sg.closed_state(StIdx::from(0 as u32)), "^", 0, 0, vec!["$"]);
-        state_exists(&grm, &sg.closed_state(StIdx::from(0 as u32)), "X", 0, 0, vec!["$"]);
-        state_exists(&grm, &sg.closed_state(StIdx::from(0 as u32)), "X", 1, 0, vec!["$"]);
-        state_exists(&grm, &sg.closed_state(StIdx::from(0 as u32)), "X", 2, 0, vec!["$"]);
-        state_exists(&grm, &sg.closed_state(StIdx::from(0 as u32)), "X", 3, 0, vec!["$"]);
-        state_exists(&grm, &sg.closed_state(StIdx::from(0 as u32)), "X", 4, 0, vec!["$"]);
-        state_exists(&grm, &sg.closed_state(StIdx::from(0 as u32)), "X", 5, 0, vec!["$"]);
+        assert_eq!(sg.closed_state(StIdx::from(0u32)).items.len(), 7);
+        state_exists(&grm, &sg.closed_state(StIdx::from(0u32)), "^", 0, 0, vec!["$"]);
+        state_exists(&grm, &sg.closed_state(StIdx::from(0u32)), "X", 0, 0, vec!["$"]);
+        state_exists(&grm, &sg.closed_state(StIdx::from(0u32)), "X", 1, 0, vec!["$"]);
+        state_exists(&grm, &sg.closed_state(StIdx::from(0u32)), "X", 2, 0, vec!["$"]);
+        state_exists(&grm, &sg.closed_state(StIdx::from(0u32)), "X", 3, 0, vec!["$"]);
+        state_exists(&grm, &sg.closed_state(StIdx::from(0u32)), "X", 4, 0, vec!["$"]);
+        state_exists(&grm, &sg.closed_state(StIdx::from(0u32)), "X", 5, 0, vec!["$"]);
 
-        let s1 = sg.edge(StIdx::from(0 as u32), Symbol::Term(grm.term_idx("a").unwrap())).unwrap();
+        let s1 = sg.edge(StIdx::from(0u32), Symbol::Term(grm.term_idx("a").unwrap())).unwrap();
         assert_eq!(sg.closed_state(s1).items.len(), 7);
         state_exists(&grm, &sg.closed_state(s1), "X", 0, 1, vec!["a", "d", "e", "$"]);
         state_exists(&grm, &sg.closed_state(s1), "X", 1, 1, vec!["a", "d", "e", "$"]);
@@ -527,7 +527,7 @@ mod test {
         state_exists(&grm, &sg.closed_state(s1), "Z", 0, 0, vec!["c"]);
         state_exists(&grm, &sg.closed_state(s1), "T", 0, 0, vec!["a", "d", "e", "$"]);
 
-        let s7 = sg.edge(StIdx::from(0 as u32), Symbol::Term(grm.term_idx("b").unwrap())).unwrap();
+        let s7 = sg.edge(StIdx::from(0u32), Symbol::Term(grm.term_idx("b").unwrap())).unwrap();
         assert_eq!(sg.closed_state(s7).items.len(), 7);
         state_exists(&grm, &sg.closed_state(s7), "X", 3, 1, vec!["a", "d", "e", "$"]);
         state_exists(&grm, &sg.closed_state(s7), "X", 4, 1, vec!["a", "d", "e", "$"]);
@@ -588,7 +588,7 @@ mod test {
         // Ommitted successors from the graph in Fig.3
 
         // X-successor of S0
-        let s0x = sg.edge(StIdx::from(0 as u32), Symbol::Nonterm(grm.nonterm_idx("X").unwrap())).unwrap();
+        let s0x = sg.edge(StIdx::from(0u32), Symbol::Nonterm(grm.nonterm_idx("X").unwrap())).unwrap();
         state_exists(&grm, &sg.closed_state(s0x), "^", 0, 1, vec!["$"]);
 
         // Y-successor of S1 (and it's d-successor)
@@ -656,7 +656,7 @@ mod test {
         let sg = pager_stategraph(&grm);
 
         // State 0
-        assert_eq!(sg.core_state(StIdx::from(0 as u32)).items.len(), 1);
-        state_exists(&grm, &sg.core_state(StIdx::from(0 as u32)), "^", 0, 0, vec!["$"]);
+        assert_eq!(sg.core_state(StIdx::from(0u32)).items.len(), 1);
+        state_exists(&grm, &sg.core_state(StIdx::from(0u32)), "^", 0, 0, vec!["$"]);
     }
 }
